couple changes, mostly just fixing warnings

+++ Main/DirectX11_Starter_2015/DirectX11_Starter/Bloom.h
@@ -5,7 +5,7 @@
 class Bloom : public PostProcessBase
 {
 public:
-	Bloom(float width, float height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler);
+	Bloom(int width, int height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler);
 	~Bloom();
 	virtual ID3D11ShaderResourceView* draw(ID3D11ShaderResourceView* ppSRV);
 private:
