+++ Main/DirectX11_Starter_2015/DirectX11_Starter/EdgeDetect.cpp
@@ -2,7 +2,7 @@
 
 
 
-EdgeDetect::EdgeDetect(float width, float height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext)
+EdgeDetect::EdgeDetect(int width, int height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext)
 {
 	device = _device;
 	deviceContext = _deviceContext;
