+++ Main/DirectX11_Starter_2015/DirectX11_Starter/FXAA.h
@@ -5,7 +5,7 @@
 class FXAA : public PostProcessBase
 {
 public:
-	FXAA(float width, float height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler);
+	FXAA(int width, int height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler);
 	virtual SRV* draw(SRV* ppSRV);
 	~FXAA();
 private:
