+++ Main/DirectX11_Starter_2015/DirectX11_Starter/Pulse.cpp
@@ -2,7 +2,7 @@
 #include "WICTextureLoader.h"
 
 
-Pulse::Pulse(float width, float height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler)
+Pulse::Pulse(int width, int height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler)
 {
 	windowWidth = width;
 	windowHeight = height;
