+++ Main/DirectX11_Starter_2015/DirectX11_Starter/ScreenDepth.h
@@ -6,7 +6,7 @@
 class ScreenDepth : public PostProcessBase
 {
 public:
-	ScreenDepth(float width, float height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler, Camera* _camera, Entity** _entities, int _numMeshes);
+	ScreenDepth(int width, int height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler, Camera* _camera, Entity** _entities, int _numMeshes);
 	virtual ID3D11ShaderResourceView* draw(ID3D11ShaderResourceView* ppSRV);
 	~ScreenDepth();
 private:
